refactor: structuring project
