+++ src/application/model/Piece.java
@@ -1,10 +1,10 @@
 package application.model;
 
+import static application.game.GameBase.TILE_SIZE;
+
 import javafx.scene.layout.StackPane;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Ellipse;
-
-import static application.GameBase.TILE_SIZE;
 
 public class Piece extends StackPane {
